Wait until player is fully online before applying queried data
Previously it was possible for the data to be received and applied in a
state where the Player wasn't fully loaded yet. The actual local player
data could've already be loaded but the player not added to the
playerlist and therefore counted as online yet. This lead to the player
data getting stored into the player.dat file after that was already
loaded.

This race condition is fixed by checking whether or not we have an
active query for the player's data which only happens when the player
fired an AsyncPlayerPreLoginEvent before.